C2C E-Commerce User’s Behaviour Analysis using SQL
Business model Customer to Customer (C2C) allows customers to do business with each other. This model is growing fast with e-commerce platforms where sellers may be required to pay some amount and buyer can buy it without paying anything. E-Commerce website brings the seller and buyer to the same platform. 

In this project, students are expected to analyze user behavior in a C2C E-Commerce business. Analyzing the user’s database will help understand the business perspective. The behavior of the users can be traced in terms of business with an exploration of the user’s database.

Analyzing the user's database will lead to understanding the business perspective. Behaviour of the users can be traced in terms of business with exploration of the user’s database. 

Dataset: One .csv file with name users_data with 98913 rows and 27 columns